Split track and rollbar methods from utility. Replace method instances in scripts.
